When building with GCC 8, there are several errors regarding packed-not-aligned.
./include/grub/gpt_partition.h:79:1: error: alignment 1 of ‘struct
grub_gpt_partentry’ is less than 8 [-Werror=packed-not-aligned]

This patch tries to fix the build error by cleaning up the ambiguity of
placing
aligned structure in a packed one. In "struct grub_btrfs_time" and
"struct
grub_gpt_part_type", the aligned attribute seems to be superfluous, and
also
has to be packed, to ensure the structure is bit-to-bit mapped to the
format
laid on disk. I think we could blame to copy and paste error here for
the
mistake. In "struct efi_variable", we have to use
grub_efi_packed_guid_t, as
the name suggests. :)

Signed-off-by: Michael Chang <address@hidden>
Tested-by: Michael Chang <address@hidden>